[refactor] 패키지명 변경에 따른 import 코드 수정
패키지명 변경: com.example.funbox -> com.rpg.funbox
