--- Android/FunBox/app/src/main/java/com/example/funbox/presentation/game/quiz/QuizBindingAdapters.kt
@@ -1,3 +1,0 @@
-package com.example.funbox.presentation.game.quiz
-
-import androidx.databinding.BindingAdapter
