--- Android/FunBox/app/src/main/java/com/example/funbox/presentation/map/MapBindingAdapters.kt
@@ -1,4 +1,0 @@
-package com.example.funbox.presentation.map
-
-class MapBindingAdapters {
-}
