--- Android/FunBox/app/src/main/java/com/example/funbox/presentation/setting/SettingBindingAdapter.kt
@@ -1,4 +1,0 @@
-package com.example.funbox.presentation.setting
-
-object SettingBindingAdapter {
-}
