+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/game/quiz/QuizBindingAdapters.kt
@@ -1,0 +1,3 @@
+package com.rpg.funbox.presentation.game.quiz
+
+import androidx.databinding.BindingAdapter
