+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/map/MapBindingAdapters.kt
@@ -1,0 +1,4 @@
+package com.rpg.funbox.presentation.map
+
+class MapBindingAdapters {
+}
