+++ Android/FunBox/app/src/main/java/com/rpg/funbox/presentation/setting/SettingBindingAdapter.kt
@@ -1,0 +1,4 @@
+package com.rpg.funbox.presentation.setting
+
+object SettingBindingAdapter {
+}
