+++ Android/FunBox/app/src/main/res/layout/fragment_map.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="vm"
-            type="com.example.funbox.presentation.map.MapViewModel" />
+            type="com.rpg.funbox.presentation.map.MapViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
