+++ Android/FunBox/app/src/main/res/layout/fragment_network_alert.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.example.funbox.presentation.game.quiz.QuizViewModel" />
+            type="com.rpg.funbox.presentation.game.quiz.QuizViewModel" />
 
     </data>
