+++ Android/FunBox/app/src/main/res/layout/fragment_score_board.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.example.funbox.presentation.game.quiz.QuizViewModel" />
+            type="com.rpg.funbox.presentation.game.quiz.QuizViewModel" />
 
     </data>
