+++ Android/FunBox/app/src/main/res/layout/fragment_setting.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="fm"
-            type="com.example.funbox.presentation.setting.SettingFragment" />
+            type="com.rpg.funbox.presentation.setting.SettingFragment" />
     </data>
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
