+++ Android/FunBox/app/src/main/res/layout/fragment_wait.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.example.funbox.presentation.game.wait.WaitViewModel" />
+            type="com.rpg.funbox.presentation.game.wait.WaitViewModel" />
 
     </data>
