+++ Android/FunBox/app/src/main/res/layout/to_be_determined_list.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="tbd"
-            type="com.example.funbox.presentation.game.gameselect.GameSelectUiState.ToBeDetermined" />
+            type="com.rpg.funbox.presentation.game.gameselect.GameSelectUiState.ToBeDetermined" />
 
     </data>
