+++ Android/FunBox/app/src/main/res/navigation/title_nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/TitleFragment"
-        android:name="com.example.funbox.presentation.login.title.TitleFragment"
+        android:name="com.rpg.funbox.presentation.login.title.TitleFragment"
         android:label="@string/fragment_title_label"
         tools:layout="@layout/fragment_title">
         <action
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/NicknameFragment"
-        android:name="com.example.funbox.presentation.login.nickname.NicknameFragment"
+        android:name="com.rpg.funbox.presentation.login.nickname.NicknameFragment"
         android:label="@string/fragment_nickname_label"
         tools:layout="@layout/fragment_nickname">
         <action
@@ -31,7 +31,7 @@
 
     <fragment
         android:id="@+id/ProfileFragment"
-        android:name="com.example.funbox.presentation.login.profile.ProfileFragment"
+        android:name="com.rpg.funbox.presentation.login.profile.ProfileFragment"
         android:label="@string/fragment_profile_label"
         tools:layout="@layout/fragment_profile">
     </fragment>
